requirements.txt
# Agent Orchestration
crewai[tools]>=0.80.0
langchain-groq>=0.2.0

# Search
tavily-python>=0.5.0

# Vector Store
chromadb>=0.5.0

# Database
supabase>=2.0.0

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.0.0

# Utilities
python-dotenv>=1.0.0
markdown>=3.5.0
pydantic>=2.0.0
